Updated DesktopPane, added Cascade, misc. edits.
Clicking a tree branch (single/double) will display its corresponding content in the right scrollpane.
Selecting a different folder in the left scrollpane causes the right scrollpane to redraw to show the folders and files that belong to the newly selected folder.
Cascade stacks the open frames in a nice, neat overlapping style.
Added documentation/comments for all methods.